Move the common functions into utils.go
